moved otpVerificationPage into  authentication folder
