+++ web/app/otp/page.tsx
@@ -1,5 +1,5 @@
 import React from "react";
-import OtpVerification from "@/components/OtpVerification";
+import OtpVerification from "@/components/Authentication/OtpVerification";
 
 const Otp: React.FC = () => {
   return (
