+++ web/components/Authentication/OtpVerification.tsx
@@ -9,13 +9,13 @@ const OtpVerification: React.FC = () => {
 
   return (
     <>
-      <div className="px-[40px] h-screen">
+      <div className="px-[40px] h-screen bg-pageColor">
         <div className="flex md:flex-row flex-col justify-between">
           <div className="h-screen w-screen flex items-center justify-center">
             <div className=" flex flex-col items-center justify-center">
               <Image src="/logo.svg" alt="logo" width={124} height={180} />
-              <h1 className="text-4xl font-bold text-[#007DFC] mt-[20px]">OTP Verification</h1>
-              <p className="text-[#3A3A3A] text-[16px] pt-[40px] pb-[30px]">
+              <h1 className="text-4xl font-bold text-btnColor mt-[20px]">OTP Verification</h1>
+              <p className="text-normalTextColor text-[16px] pt-[40px] pb-[30px]">
                 Please enter the OTP sent to your email
               </p>
               <OtpInput
@@ -35,16 +35,18 @@ const OtpVerification: React.FC = () => {
                   <input
                     {...props}
                     type="password"
-                    className="bg-[#605DEC14] outline-[0.9px] outline-[#0000001A] text-[14px]"
+                    className="bg-inputColor outline-[0.9px] outline-[#0000001A] text-[14px]"
                   />
                 )}
               />
               <div className="pt-7 pb-10">
                 <p className="text-[#3A3A3A80]">
-                  Didn&apos;t receive OTP? <span className="text-[#007DFC] underline">Resend</span>
+                  Didn&apos;t receive OTP? <span className="text-btnColor underline">Resend</span>
                 </p>
               </div>
-              <button className="bg-[#007DFC] text-white px-12 py-3 rounded-[24px]">Submit</button>
+              <button className="bg-btnColor text-btnTextColor px-12 py-3 rounded-[24px]">
+                Submit
+              </button>
             </div>
           </div>
